WIP rename fpan-filter to site-filter; force addition to search url
