+++ fpan/templates/views/components/search/site-filter.htm
@@ -1,0 +1,1 @@
+<div>Your access to Archaeological Sites is restricted.</div>
